Added a new upper_triangle class to handle sparse matrices where only the upper half of the square matrix is used.
Changed the RuntimeDB::interactions matrix into an upper_triangle instance.
This instance uses the symmetry fixing.

Using the upper_triangle class, it is trivial to use an iterator to iterate
through the various elements in the half-full upper triangular matrix.  

RuntimeDB::props (particle properties vector) is now sorted at the time of
initInteractions().  The order is by mass, lowest to highest.  Added several
notes and small changes to underline the fact that the order of particle
properties is ill-determined until AFTER initInteractions() is called.  



git-svn-id: file:///home/olsonse/src/svn-repository/dsmc/branches/plasma-fieldsolve/src/particledb@626 e877f351-6dfb-0310-a246-97bd42f90100