Removed interaction::Output class and moved some pieces into Equation.
Other pieces were moved into the appropriate location.  In the future, I'd like
to remove the Input class and use the same code to represent inputs as outputs.
This would allow >2 inputs and would provide a more generic interface to the
left hand side and right hand side of the interaction equations.