Cross sections moved into new namespace and are loaded using a Input reference.
This is the same loading signature that the interaction models use.  Perhaps in
the future, the loading functions should be passed a reference to the full
Equation object, but not yet...
These changes had the following consequences:
reduced mass is no longer stored by the cross_section::Base class, but must be
computed and stored in the descendant classes.  This makes for a more generic
interface for cross sections and their specific loading functions.

Also, cross_section::Base classes now have the 'options' template parameter,
similar to RuntimeDB and interaction::model::Base classes.

The ReducedMass class was moved from a detail namespace to the interaction
namespace since it will be used by several pieces and could be useful to the
outside world.

Created new interact(...) signature for arbitrary numbers of interaction terms
for:  InElastic and VSSElastic.