Cleaned master branch: rm bolsig+ conversion and testdist.C
The bolsig+ conversion is in the form of VIM search/replace regex commands.
This will be left for now in my own private imports/bolsig branch.

testdist.C is some code for developing the precomputed set (as of yet
unfinished).  This is also left in my private branch for the day that I finish
that feature.