Extrapolation made a bit better.
Errors are emitted (exceptions thrown) when the data cannot be reasonably
extrapolated as ln(E)/E.
Warning are emitted the first time data is requested that requires
extrapolations and for which extrapolations can be reasonably done.  A count of
the number of extrapolations is kept for use at a later time.