Data that goes to zero doesn't get extrapolated.
Data that goes to zero, stays at zero and doesn't issue a "beyond the end of
data" error or an extrapolation warning.

Also fixed the order of the extrapolation anchor points (the last three points
that were used for extrapolating were given in the wrong order according to what
I had planned).

CrossSection tests are modified so that they write out the correct number of
lines so-as to not confuse the plotting script.  The plotting script was
modified to show which lines of how many total lines are currently being plotted
(useful for when the plots=? command line option is given).